+++ metadata.go
@@ -39,15 +39,13 @@ import (
 
 // schema metadata for a keyspace
 type KeyspaceMetadata struct {
-	Name            string
-	DurableWrites   bool
-	StrategyClass   string
-	StrategyOptions map[string]interface{}
-	Tables          map[string]*TableMetadata
-	Functions       map[string]*FunctionMetadata
-	Aggregates      map[string]*AggregateMetadata
-	// Deprecated: use the MaterializedViews field for views and UserTypes field for udts instead.
-	Views             map[string]*ViewMetadata
+	Name              string
+	DurableWrites     bool
+	StrategyClass     string
+	StrategyOptions   map[string]interface{}
+	Tables            map[string]*TableMetadata
+	Functions         map[string]*FunctionMetadata
+	Aggregates        map[string]*AggregateMetadata
 	MaterializedViews map[string]*MaterializedViewMetadata
 	UserTypes         map[string]*UserTypeMetadata
 }
@@ -109,19 +107,11 @@ type AggregateMetadata struct {
 	finalFunc string
 }
 
-// ViewMetadata holds the metadata for views.
-// Deprecated: this is kept for backwards compatibility issues. Use MaterializedViewMetadata.
-type ViewMetadata struct {
-	Keyspace   string
-	Name       string
-	FieldNames []string
-	FieldTypes []TypeInfo
-}
-
 // MaterializedViewMetadata holds the metadata for materialized views.
 type MaterializedViewMetadata struct {
 	Keyspace                string
 	Name                    string
+	AdditionalWritePolicy   string
 	BaseTableId             UUID
 	BaseTable               *TableMetadata
 	BloomFilterFpChance     float64
@@ -139,7 +129,8 @@ type MaterializedViewMetadata struct {
 	MaxIndexInterval        int
 	MemtableFlushPeriodInMs int
 	MinIndexInterval        int
-	ReadRepairChance        float64
+	ReadRepair              string  // Only present in Cassandra 4.0+
+	ReadRepairChance        float64 // Note: Cassandra 4.0 removed ReadRepairChance and added ReadRepair instead
 	SpeculativeRetry        string
 
 	baseTableName string
@@ -304,7 +295,7 @@ func (s *schemaDescriber) refreshSchema(keyspaceName string) error {
 	if err != nil {
 		return err
 	}
-	views, err := getViewsMetadata(s.session, keyspaceName)
+	userTypes, err := getUserTypeMetadata(s.session, keyspaceName)
 	if err != nil {
 		return err
 	}
@@ -314,7 +305,7 @@ func (s *schemaDescriber) refreshSchema(keyspaceName string) error {
 	}
 
 	// organize the schema data
-	compileMetadata(s.session.cfg.ProtoVersion, keyspace, tables, columns, functions, aggregates, views,
+	compileMetadata(s.session.cfg.ProtoVersion, keyspace, tables, columns, functions, aggregates, userTypes,
 		materializedViews, s.session.logger)
 
 	// update the cache
@@ -335,7 +326,7 @@ func compileMetadata(
 	columns []ColumnMetadata,
 	functions []FunctionMetadata,
 	aggregates []AggregateMetadata,
-	views []ViewMetadata,
+	uTypes []UserTypeMetadata,
 	materializedViews []MaterializedViewMetadata,
 	logger StdLogger,
 ) {
@@ -355,22 +346,9 @@ func compileMetadata(
 		aggregates[i].StateFunc = *keyspace.Functions[aggregates[i].stateFunc]
 		keyspace.Aggregates[aggregates[i].Name] = &aggregates[i]
 	}
-	keyspace.Views = make(map[string]*ViewMetadata, len(views))
-	for i := range views {
-		keyspace.Views[views[i].Name] = &views[i]
-	}
-	// Views currently holds the types and hasn't been deleted for backward compatibility issues.
-	// That's why it's ok to copy Views into Types in this case. For the real Views use MaterializedViews.
-	types := make([]UserTypeMetadata, len(views))
-	for i := range views {
-		types[i].Keyspace = views[i].Keyspace
-		types[i].Name = views[i].Name
-		types[i].FieldNames = views[i].FieldNames
-		types[i].FieldTypes = views[i].FieldTypes
-	}
-	keyspace.UserTypes = make(map[string]*UserTypeMetadata, len(views))
-	for i := range types {
-		keyspace.UserTypes[types[i].Name] = &types[i]
+	keyspace.UserTypes = make(map[string]*UserTypeMetadata, len(uTypes))
+	for i := range uTypes {
+		keyspace.UserTypes[uTypes[i].Name] = &uTypes[i]
 	}
 	keyspace.MaterializedViews = make(map[string]*MaterializedViewMetadata, len(materializedViews))
 	for i, _ := range materializedViews {
@@ -954,7 +932,7 @@ func getTypeInfo(t string, logger StdLogger) TypeInfo {
 	return getCassandraType(t, logger)
 }
 
-func getViewsMetadata(session *Session, keyspaceName string) ([]ViewMetadata, error) {
+func getUserTypeMetadata(session *Session, keyspaceName string) ([]UserTypeMetadata, error) {
 	if session.cfg.ProtoVersion == protoVersion1 {
 		return nil, nil
 	}
@@ -972,31 +950,210 @@ func getViewsMetadata(session *Session, keyspaceName string) ([]ViewMetadata, er
 		FROM %s
 		WHERE keyspace_name = ?`, tableName)
 
-	var views []ViewMetadata
+	var uTypes []UserTypeMetadata
 
 	rows := session.control.query(stmt, keyspaceName).Scanner()
 	for rows.Next() {
-		view := ViewMetadata{Keyspace: keyspaceName}
+		uType := UserTypeMetadata{Keyspace: keyspaceName}
 		var argumentTypes []string
-		err := rows.Scan(&view.Name,
-			&view.FieldNames,
+		err := rows.Scan(&uType.Name,
+			&uType.FieldNames,
 			&argumentTypes,
 		)
 		if err != nil {
 			return nil, err
 		}
-		view.FieldTypes = make([]TypeInfo, len(argumentTypes))
+		uType.FieldTypes = make([]TypeInfo, len(argumentTypes))
 		for i, argumentType := range argumentTypes {
-			view.FieldTypes[i] = getTypeInfo(argumentType, session.logger)
+			uType.FieldTypes[i] = getTypeInfo(argumentType, session.logger)
 		}
-		views = append(views, view)
+		uTypes = append(uTypes, uType)
 	}
 
 	if err := rows.Err(); err != nil {
 		return nil, err
 	}
 
-	return views, nil
+	return uTypes, nil
+}
+
+func bytesMapToStringsMap(byteData map[string][]byte) map[string]string {
+	extensions := make(map[string]string, len(byteData))
+	for key, rowByte := range byteData {
+		extensions[key] = string(rowByte)
+	}
+
+	return extensions
+}
+
+func materializedViewMetadataFromMap(currentObject map[string]interface{}, materializedView *MaterializedViewMetadata) error {
+	const errorMessage = "gocql.materializedViewMetadataFromMap failed to read column %s"
+	var ok bool
+	for key, value := range currentObject {
+		switch key {
+		case "keyspace_name":
+			materializedView.Keyspace, ok = value.(string)
+			if !ok {
+				return fmt.Errorf(errorMessage, key)
+			}
+
+		case "view_name":
+			materializedView.Name, ok = value.(string)
+			if !ok {
+				return fmt.Errorf(errorMessage, key)
+			}
+
+		case "additional_write_policy":
+			materializedView.AdditionalWritePolicy, ok = value.(string)
+			if !ok {
+				return fmt.Errorf(errorMessage, key)
+			}
+
+		case "base_table_id":
+			materializedView.BaseTableId, ok = value.(UUID)
+			if !ok {
+				return fmt.Errorf(errorMessage, key)
+			}
+
+		case "base_table_name":
+			materializedView.baseTableName, ok = value.(string)
+			if !ok {
+				return fmt.Errorf(errorMessage, key)
+			}
+
+		case "bloom_filter_fp_chance":
+			materializedView.BloomFilterFpChance, ok = value.(float64)
+			if !ok {
+				return fmt.Errorf(errorMessage, key)
+			}
+
+		case "caching":
+			materializedView.Caching, ok = value.(map[string]string)
+			if !ok {
+				return fmt.Errorf(errorMessage, key)
+			}
+
+		case "comment":
+			materializedView.Comment, ok = value.(string)
+			if !ok {
+				return fmt.Errorf(errorMessage, key)
+			}
+
+		case "compaction":
+			materializedView.Compaction, ok = value.(map[string]string)
+			if !ok {
+				return fmt.Errorf(errorMessage, key)
+			}
+
+		case "compression":
+			materializedView.Compression, ok = value.(map[string]string)
+			if !ok {
+				return fmt.Errorf(errorMessage, key)
+			}
+
+		case "crc_check_chance":
+			materializedView.CrcCheckChance, ok = value.(float64)
+			if !ok {
+				return fmt.Errorf(errorMessage, key)
+			}
+
+		case "dclocal_read_repair_chance":
+			materializedView.DcLocalReadRepairChance, ok = value.(float64)
+			if !ok {
+				return fmt.Errorf(errorMessage, key)
+			}
+
+		case "default_time_to_live":
+			materializedView.DefaultTimeToLive, ok = value.(int)
+			if !ok {
+				return fmt.Errorf(errorMessage, key)
+			}
+
+		case "extensions":
+			byteData, ok := value.(map[string][]byte)
+			if !ok {
+				return fmt.Errorf(errorMessage, key)
+			}
+
+			materializedView.Extensions = bytesMapToStringsMap(byteData)
+
+		case "gc_grace_seconds":
+			materializedView.GcGraceSeconds, ok = value.(int)
+			if !ok {
+				return fmt.Errorf(errorMessage, key)
+			}
+
+		case "id":
+			materializedView.Id, ok = value.(UUID)
+			if !ok {
+				return fmt.Errorf(errorMessage, key)
+			}
+
+		case "include_all_columns":
+			materializedView.IncludeAllColumns, ok = value.(bool)
+			if !ok {
+				return fmt.Errorf(errorMessage, key)
+			}
+
+		case "max_index_interval":
+			materializedView.MaxIndexInterval, ok = value.(int)
+			if !ok {
+				return fmt.Errorf(errorMessage, key)
+			}
+
+		case "memtable_flush_period_in_ms":
+			materializedView.MemtableFlushPeriodInMs, ok = value.(int)
+			if !ok {
+				return fmt.Errorf(errorMessage, key)
+			}
+
+		case "min_index_interval":
+			materializedView.MinIndexInterval, ok = value.(int)
+			if !ok {
+				return fmt.Errorf(errorMessage, key)
+			}
+
+		case "read_repair":
+			materializedView.ReadRepair, ok = value.(string)
+			if !ok {
+				return fmt.Errorf(errorMessage, key)
+			}
+
+		case "read_repair_chance":
+			materializedView.ReadRepairChance, ok = value.(float64)
+			if !ok {
+				return fmt.Errorf(errorMessage, key)
+			}
+
+		case "speculative_retry":
+			materializedView.SpeculativeRetry, ok = value.(string)
+			if !ok {
+				return fmt.Errorf(errorMessage, key)
+			}
+
+		}
+	}
+	return nil
+}
+
+func parseSystemSchemaViews(iter *Iter) ([]MaterializedViewMetadata, error) {
+	var materializedViews []MaterializedViewMetadata
+	s, err := iter.SliceMap()
+	if err != nil {
+		return nil, err
+	}
+
+	for _, row := range s {
+		var materializedView MaterializedViewMetadata
+		err = materializedViewMetadataFromMap(row, &materializedView)
+		if err != nil {
+			return nil, err
+		}
+
+		materializedViews = append(materializedViews, materializedView)
+	}
+
+	return materializedViews, nil
 }
 
 func getMaterializedViewsMetadata(session *Session, keyspaceName string) ([]MaterializedViewMetadata, error) {
@@ -1005,63 +1162,16 @@ func getMaterializedViewsMetadata(session *Session, keyspaceName string) ([]Mate
 	}
 	var tableName = "system_schema.views"
 	stmt := fmt.Sprintf(`
-		SELECT
-			view_name,
-			base_table_id,
-			base_table_name,
-			bloom_filter_fp_chance,
-			caching,
-			comment,
-			compaction,
-			compression,
-			crc_check_chance,
-			dclocal_read_repair_chance,
-			default_time_to_live,
-			extensions,
-			gc_grace_seconds,
-			id,
-			include_all_columns,
-			max_index_interval,
-			memtable_flush_period_in_ms,
-			min_index_interval,
-			read_repair_chance,
-			speculative_retry
+		SELECT *
 		FROM %s
 		WHERE keyspace_name = ?`, tableName)
 
 	var materializedViews []MaterializedViewMetadata
 
-	rows := session.control.query(stmt, keyspaceName).Scanner()
-	for rows.Next() {
-		materializedView := MaterializedViewMetadata{Keyspace: keyspaceName}
-		err := rows.Scan(&materializedView.Name,
-			&materializedView.BaseTableId,
-			&materializedView.baseTableName,
-			&materializedView.BloomFilterFpChance,
-			&materializedView.Caching,
-			&materializedView.Comment,
-			&materializedView.Compaction,
-			&materializedView.Compression,
-			&materializedView.CrcCheckChance,
-			&materializedView.DcLocalReadRepairChance,
-			&materializedView.DefaultTimeToLive,
-			&materializedView.Extensions,
-			&materializedView.GcGraceSeconds,
-			&materializedView.Id,
-			&materializedView.IncludeAllColumns,
-			&materializedView.MaxIndexInterval,
-			&materializedView.MemtableFlushPeriodInMs,
-			&materializedView.MinIndexInterval,
-			&materializedView.ReadRepairChance,
-			&materializedView.SpeculativeRetry,
-		)
-		if err != nil {
-			return nil, err
-		}
-		materializedViews = append(materializedViews, materializedView)
-	}
+	iter := session.control.query(stmt, keyspaceName)
 
-	if err := rows.Err(); err != nil {
+	materializedViews, err := parseSystemSchemaViews(iter)
+	if err != nil {
 		return nil, err
 	}